data_ingestion/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pytz==2023.3
notion-client==2.2.1
python-dotenv==1.0.0